% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funlog.R
\name{makeLogger}
\alias{makeLogger}
\title{Create a logging function}
\usage{
makeLogger(default_level = "trace")
}
\arguments{
\item{default_level}{The default logging level of the function. Use one of the options in the global LEVEL variable.}
}
\value{
A function taking the a message and optionally a log level
The message can use the glue libraries {} syntax.
}
\description{
Create a logging function
}
\examples{
LOG <- makeLogger(funlog::LEVEL$TRACE)
LOG("2+2={2+2}")
}
