% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funlog.R
\name{setLogLevel}
\alias{setLogLevel}
\title{Set the log level}
\usage{
setLogLevel(level)
}
\arguments{
\item{level}{Can be anything in \code{LEVEL}}
}
\description{
Set the log level
}
\examples{
setLogLevel(funlog::LEVEL$TRACE)
}
